Fix services page issues and add skeleton loading back
Now you are able to directly go to boompow.banano.cc/services without being redirected back to the main page if it's not loaded yet.
